requirements.txt
6==6
absl-py==1.4.0
aniso8601==7.0.0
asgiref==3.7.2
astroid==2.15.5
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.3
dateparser==1.1.8
decouple==0.0.7
defusedxml==0.7.1
dill==0.3.6
dj-database-url==2.1.0
Django==4.2.2
django-admin-argon-dashboard==1.0.16
django-ckeditor==6.7.0
django-js-asset==2.1.0
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
idna==3.4
isort==5.12.0
jax==0.4.10
jet-bridge-base==1.8.10
keras==2.12.0
lazy-object-proxy==1.9.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
mccabe==0.7.0
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.0
platformdirs==3.5.3
promise==2.3
prompt-toolkit==2.0.9
protobuf==4.23.0
psutil==5.9.5
psycopg==3.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.4
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
regex==2023.8.8
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.3
scipy==1.10.1
six==1.16.0
social-auth-core==4.4.2
SQLAlchemy==1.4.49
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tomlkit==0.11.8
tornado==6.3.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
